added data-format restructuring
